% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDataOverview.R
\name{PlotVarBreakdown}
\alias{PlotVarBreakdown}
\title{Plots variant frequency data as a stacked bar plot that shows the variant
breakdown for each timepoint}
\usage{
PlotVarBreakdown(varData = NA)
}
\arguments{
\item{varData}{A dataframe containing variant frequency data with the
columns 'date', 'variant', 'proportion', and optionally 'parent'}
}
\value{
Returns a plot of the variant breakdown over time.
}
\description{
A function that plots variant frequency data from a pre-processed dataframe and
titles the plot with the time frame of the data points.
}
\examples{
plot <- PlotVarBreakdown(VariantDF)
plot

}
\references{
Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4. https://ggplot2.tidyverse.org
}
