% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDataOverview.R
\name{PlotViralSignal}
\alias{PlotViralSignal}
\title{Plots viral signal data as a line plot with dots marking each signal entry}
\usage{
PlotViralSignal(signalData = NA)
}
\arguments{
\item{signalData}{A dataframe containing viral signal data with columns
'date' and 'signal'}
}
\value{
Returns a plot of the viral signal data over time.
}
\description{
A function that plots viral signal data from a pre-processed dataframe and
titles the plot with the time frame of the data points.
}
\examples{
plot <- PlotViralSignal(ViralSignalDF)
plot

}
\references{
Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4. https://ggplot2.tidyverse.org
}
