% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCovidWasteWatch.R
\name{runCovidWasteWatch}
\alias{runCovidWasteWatch}
\title{Launch Shiny App for CovidWasteWatch}
\usage{
runCovidWasteWatch()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for CovidWasteWatch.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
  CovidWasteWatch::CovidWasteWatch()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
