% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarBreakdown.R
\name{VarBreakdown}
\alias{VarBreakdown}
\title{Loads and provides overview of variant frequency data}
\usage{
VarBreakdown(fileVariant = NA, parentVariants = TRUE)
}
\arguments{
\item{fileVariant}{A name/path of the variant data file}

\item{parentVariants}{A Boolean indicating whether the data contains a
parent variants column.}
}
\value{
Returns a results list containing the number of unique variants
and, if parent column exists, the list of variants grouped by parents.
}
\description{
A function that extracts relevant data from the file, calculates the number of
unique variants and groups variants by parents, and plots the variant
breakdown over time.
}
\examples{
filepath <- system.file("extdata", "variant_input_data.csv", package = "CovidWasteWatch")
results <- VarBreakdown(fileVariant = filepath)
results

}
\references{
Bengtsson, H (2022). R.utils: Various Programming Utilities.
https://henrikbengtsson.github.io/R.utils/

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4. https://ggplot2.tidyverse.org

Wickham H, François R, Henry L, Müller K, Vaughan D. (2023).
dplyr: A Grammar of Data Manipulation. https://dplyr.tidyverse.org
}
