% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VariantDF}
\alias{VariantDF}
\title{COVID-19 Viral Variant Breakdown Dataframe}
\format{
A dataframe with three columns:
\describe{
\item{date}{The date of samples in YYYY-MM-DD format.}
\item{variant}{The name of the variant detected.}
\item{parent}{The name of the parent variant of the detected variant.}
\item{proportion}{The proportion of the variant in the weekly samples, out of 1.}
}
}
\source{
The dataset was downloaded from Health Canada's website
https://health-infobase.canada.ca/covid-19/testing-variants.html
which is updated weekly. The dataset included in this package contains data
from 2023-08-27 to 2023-10-29.
}
\usage{
VariantDF
}
\description{
A dataframe of weekly variant proportion data in Canada.
This is the expected resulting dataframe when "inst/extdata/variant_input_data.csv"
is used as the input file for function VarBreakdown.
}
\examples{
\dontrun{
 VariantDF
 }
}
\keyword{datasets}
