% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputDataOverview.R
\name{ViralSignal}
\alias{ViralSignal}
\title{Loads and provides overview of viral signal data}
\usage{
ViralSignal(fileSignal = NA)
}
\arguments{
\item{fileSignal}{A name/path of the viral signal data file}
}
\value{
Returns a results list containing the number of timepoints,
average rate of change per day, and a plot of the data.
}
\description{
A function that extracts relevant data from the file, calculates the number of
unique timepoints and average rate of change per day, and plots the viral
signal over time.
}
\examples{
filepath <- system.file("extdata", "viral_signal_input_data.csv", package = "CovidWasteWatch")
results <- ViralSignal(fileSignal = filepath)
results

}
\references{
Bengtsson, H (2022). R.utils: Various Programming Utilities.
https://henrikbengtsson.github.io/R.utils/

Wickham, H. (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4. https://ggplot2.tidyverse.org

Wickham H, François R, Henry L, Müller K, Vaughan D. (2023).
dplyr: A Grammar of Data Manipulation. https://dplyr.tidyverse.org
}
