% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ViralSignalDF}
\alias{ViralSignalDF}
\title{COVID-19 Wastewater Viral Signal Dataframe}
\format{
A dataframe with two columns:
\describe{
\item{date}{The date of samples in YYYY-MM-DD format.}
\item{signal}{The signal level detected from samples.}
}
}
\source{
The dataset was downloaded from Public Health Ontario's website
https://www.publichealthontario.ca/en/Data-and-Analysis/Infectious-Disease/COVID-19-Data-Surveillance/Wastewater
which is updated weekly. The data included in this package contains data from
2022-10-26 to 2023-10-28.
}
\usage{
ViralSignalDF
}
\description{
A dataframe of viral signal data from COVID-19 wastewater surveillance in Ontario.
This is the expected resulting dataframe when "inst/extdata/viral_signal_input_data.csv"
is used as the input file for function ViralSignal.
}
\examples{
\dontrun{
 ViralSignalDF
}
}
\keyword{datasets}
